+++ js/consts.js
@@ -1,0 +1,39 @@
+const MAX_COUNT_PHOTOS = 25;
+const COUNT_COMMENTS = 8;
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const DESCRIPTION = [
+  'Красота!',
+  'Необыкновенно!',
+  'Жизнь прекрасна!',
+];
+
+const NAMES = [
+  'Анастасия',
+  'Михаил',
+  'Елизавета',
+  'Леонид',
+  'Максим',
+  'Светлана',
+];
+
+const CountLike = {
+  MIN: 15,
+  MAX: 200,
+};
+
+const CountAvatar = {
+  MIN: 1,
+  MAX: 6,
+};
+
+export {MAX_COUNT_PHOTOS, MESSAGE,DESCRIPTION, NAMES, CountAvatar, CountLike, COUNT_COMMENTS};
+
